chore: move rspack chain helpers to core
