+++ packages/core/tests/external.test.ts
@@ -1,4 +1,4 @@
-import { getBundlerChain } from '@rsbuild/shared';
+import { getBundlerChain } from '../src/configChain';
 import { pluginExternals } from '../src/plugins/externals';
 
 describe('plugin-external', () => {
