+++ packages/shared/src/chain.ts
@@ -1,56 +1,9 @@
-import type { EntryDescription } from '@rspack/core';
 import RspackChain from '../compiled/rspack-chain/index.js';
 import { NODE_MODULES_REGEX, TS_AND_JSX_REGEX } from './constants';
-import { debug } from './logger';
-import type {
-  CreateAsyncHook,
-  ModifyBundlerChainFn,
-  ModifyBundlerChainUtils,
-  NormalizedConfig,
-  RsbuildConfig,
-  RsbuildContext,
-  RsbuildEntry,
-  RspackConfig,
-} from './types';
-import { isPlainObject } from './utils';
+import type { NormalizedConfig, RsbuildContext } from './types';
 import { castArray } from './utils';
 
 export { RspackChain };
-
-export async function getBundlerChain() {
-  const bundlerChain = new RspackChain();
-
-  return bundlerChain as unknown as RspackChain;
-}
-
-export async function modifyBundlerChain(
-  context: RsbuildContext & {
-    hooks: {
-      modifyBundlerChain: CreateAsyncHook<ModifyBundlerChainFn>;
-    };
-    config: Readonly<RsbuildConfig>;
-  },
-  utils: ModifyBundlerChainUtils,
-): Promise<RspackChain> {
-  debug('modify bundler chain');
-
-  const bundlerChain = await getBundlerChain();
-
-  const [modifiedBundlerChain] = await context.hooks.modifyBundlerChain.call(
-    bundlerChain,
-    utils,
-  );
-
-  if (context.config.tools?.bundlerChain) {
-    for (const item of castArray(context.config.tools.bundlerChain)) {
-      await item(modifiedBundlerChain, utils);
-    }
-  }
-
-  debug('modify bundler chain done');
-
-  return modifiedBundlerChain;
-}
 
 export const CHAIN_ID = {
   /** Predefined rules */
@@ -264,53 +217,4 @@ export function applyScriptCondition({
   for (const condition of [...excludes, ...(config.source.exclude || [])]) {
     rule.exclude.add(condition);
   }
-}
-
-export function chainToConfig(chain: RspackChain): RspackConfig {
-  const config = chain.toConfig();
-  const { entry } = config;
-
-  if (!isPlainObject(entry)) {
-    return config as RspackConfig;
-  }
-
-  const formattedEntry: RsbuildEntry = {};
-
-  /**
-   * rspack-chain can not handle entry description object correctly,
-   * so we need to format the entry object and correct the entry description object.
-   */
-  for (const [entryName, entryValue] of Object.entries(entry)) {
-    const entryImport: string[] = [];
-    let entryDescription: EntryDescription | null = null;
-
-    for (const item of castArray(entryValue)) {
-      if (typeof item === 'string') {
-        entryImport.push(item);
-        continue;
-      }
-
-      if (item.import) {
-        entryImport.push(...castArray(item.import));
-      }
-
-      if (entryDescription) {
-        // merge entry description object
-        Object.assign(entryDescription, item);
-      } else {
-        entryDescription = item;
-      }
-    }
-
-    formattedEntry[entryName] = entryDescription
-      ? {
-          ...(entryDescription as EntryDescription),
-          import: entryImport,
-        }
-      : entryImport;
-  }
-
-  config.entry = formattedEntry;
-
-  return config as RspackConfig;
 }
